fix: Gate AI Counselor behind premium and fix Decimal test assertions
Premium gating:
- Replace NavigationLink with Button that checks appState.isPremium
- Non-premium users see the paywall sheet instead of the chat
- Premium users get the counselor in a sheet with NavigationStack
- Show chevron for premium users, "Premium" badge for others

Test alignment:
- Convert Decimal to .doubleValue before XCTAssertEqual with accuracy
  parameter (Decimal doesn't conform to FloatingPoint)

+++ TitheStewardTests/GivingServiceIntegrationTests.swift
@@ -57,7 +57,7 @@ final class GivingServiceIntegrationTests: XCTestCase {
 
         let total = service.monthlyRecurringTotal()
         // 100 * 4.33 = 433
-        XCTAssertEqual(total, 433, accuracy: 1)
+        XCTAssertEqual(total.doubleValue, 433, accuracy: 1)
     }
 
     func testMonthlyRecurringTotalBiweekly() {
@@ -69,7 +69,7 @@ final class GivingServiceIntegrationTests: XCTestCase {
 
         let total = service.monthlyRecurringTotal()
         // 200 * 2.17 = 434
-        XCTAssertEqual(total, 434, accuracy: 1)
+        XCTAssertEqual(total.doubleValue, 434, accuracy: 1)
     }
 
     func testMonthlyRecurringTotalQuarterly() {
